🧪 ZMVZRZ cleanup: split remote cleanup regression
