+++ .agentplane/tasks/202607252051-ZMVZRZ/pr/diffstat.txt
@@ -1,5 +1,5 @@
- .../run-cli.core.pr-flow.cleanup-merged.test.ts    | 137 ++++++++++++++++++++
- .../src/commands/branch/cleanup-merged.ts          |  17 ++-
- .../commands/shared/merged-branch-cleanup.test.ts  | 144 ++++++++++++++++++++-
- .../src/commands/shared/merged-branch-cleanup.ts   |  85 +++++++++++-
- 4 files changed, 376 insertions(+), 7 deletions(-)
+ ...-cli.core.pr-flow.cleanup-merged.remote.test.ts | 171 +++++++++++++++++++++
+ .../src/commands/branch/cleanup-merged.ts          |  17 +-
+ .../commands/shared/merged-branch-cleanup.test.ts  | 144 ++++++++++++++++-
+ .../src/commands/shared/merged-branch-cleanup.ts   |  85 +++++++++-
+ 4 files changed, 410 insertions(+), 7 deletions(-)
