+++ .agentplane/tasks/202607252051-ZMVZRZ/pr/github-title.txt
@@ -1,1 +1,1 @@
-✅ ZMVZRZ task: Make merged worktree cleanup resilient to partial removal [202607252051-ZMVZRZ]
+🚧 ZMVZRZ task: Make merged worktree cleanup resilient to partial removal [202607252051-ZMVZRZ]
