+++ .agentplane/tasks/202607252051-ZMVZRZ/pr/meta.json
@@ -2,23 +2,16 @@
   "base": "main",
   "branch": "task/202607252051-ZMVZRZ/make-merged-worktree-cleanup-resilient",
   "created_at": "2026-07-25T22:13:26.625Z",
-  "diffstat_sha256": "sha256:6f9f8ab37947b2914bad88368bf4d6d7a9223629af90825ee540967512480f97",
-  "last_verified_at": "2026-07-25T22:45:07.482Z",
+  "diffstat_sha256": "sha256:a7773780b203840da8ef63609e8f7dbeed72fc4be5be0ef602bac49f21b8f2bf",
+  "last_verified_at": "2026-07-25T23:13:40.113Z",
   "last_verified_diffstat_sha256": "sha256:6f9f8ab37947b2914bad88368bf4d6d7a9223629af90825ee540967512480f97",
   "pr_number": 4622,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/4622",
-  "pre_merge_closure": {
-    "basis_commit": "6c19d647ac0428ba1ffe5dbec72e87a0ad75b0b3",
-    "branch": "task/202607252051-ZMVZRZ/make-merged-worktree-cleanup-resilient",
-    "pr_number": 4622,
-    "recorded_at": "2026-07-25T22:48:33.920Z",
-    "state": "closed_before_merge"
-  },
   "schema_version": 1,
   "status": "OPEN",
   "task_id": "202607252051-ZMVZRZ",
   "updated_at": "2026-07-25T22:14:25.745Z",
   "verify": {
-    "status": "pass"
+    "status": "fail"
   }
 }
