+++ .agentplane/tasks/202607252051-ZMVZRZ/pr/review.md
@@ -6,14 +6,14 @@ Created: 2026-07-25T22:13:26.625Z
 
 - Task: `202607252051-ZMVZRZ`
 - Title: Make merged worktree cleanup resilient to partial removal
-- Status: DONE
+- Status: DOING
 - Branch: `task/202607252051-ZMVZRZ/make-merged-worktree-cleanup-resilient`
 - Canonical task record: `.agentplane/tasks/202607252051-ZMVZRZ/README.md`
 
 ## Verification
 
-- State: ok
-- Note: Independent TESTER verification passed for PR #4622 at 6c19d647.
+- State: needs_rework
+- Note: Hosted Core CI #30178341063 failed verify-contract at the hotspot baseline; task requires a bounded test split before re-verification.
 - Canonical workflow state lives in the task README.
 
 ## Handoff Notes
@@ -29,11 +29,11 @@ Created: 2026-07-25T22:13:26.625Z
 - Head: computed live by `agentplane pr check` / `agentplane integrate`
 
 ```text
- .../run-cli.core.pr-flow.cleanup-merged.test.ts    | 137 ++++++++++++++++++++
- .../src/commands/branch/cleanup-merged.ts          |  17 ++-
- .../commands/shared/merged-branch-cleanup.test.ts  | 144 ++++++++++++++++++++-
- .../src/commands/shared/merged-branch-cleanup.ts   |  85 +++++++++++-
- 4 files changed, 376 insertions(+), 7 deletions(-)
+ ...-cli.core.pr-flow.cleanup-merged.remote.test.ts | 171 +++++++++++++++++++++
+ .../src/commands/branch/cleanup-merged.ts          |  17 +-
+ .../commands/shared/merged-branch-cleanup.test.ts  | 144 ++++++++++++++++-
+ .../src/commands/shared/merged-branch-cleanup.ts   |  85 +++++++++-
+ 4 files changed, 410 insertions(+), 7 deletions(-)
 ```
 
 </details>
